+++ package.json
@@ -28,7 +28,7 @@
     "test": "npm run unit",
     "lint": "eslint --ext .js,.vue src test/unit",
     "build": "node build/build.js",
-    "build-vga": "mkdir -p ./dist-vga && babel ./src/components/vga/ -d ./dist-vga"
+    "build-vga": "mkdir -p ./dist-vga && babel ./src/vga/ -d ./dist-vga"
   },
   "dependencies": {
     "vue": "^2.5.2",
